+++ src/App.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 /* eslint-disable jsx-a11y/control-has-associated-label */
 
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import { Todo } from './types/Todo';
 import * as todoService from './api/todos';
@@ -11,7 +11,6 @@ import { Footer } from './components/Footer/Footer';
 import { Header } from './components/Header/Header';
 // eslint-disable-next-line
 import { ErrorNotification } from './components/ErrorNotification/ErrorNotification';
-import { TodoList } from './components/Todolist/TodoList';
 import { UserWarning } from './UserWarning';
 
 const ERROR_DELAY = 3000;
@@ -34,7 +33,7 @@ export const App: React.FC = () => {
     if (!isDisabledInput && inputRef.current && loadingIds) {
       inputRef.current.focus();
     }
-  }, [isDisabledInput, loadingIds, errorMessage]);
+  }, [isDisabledInput, loadingIds]);
 
   useEffect(() => {
     todoService
@@ -220,6 +219,19 @@ export const App: React.FC = () => {
     return `${leftItems} ${leftItems === 1 ? 'item' : 'items'} left`;
   };
 
+  const filteredTodos = useMemo(() => {
+    switch (filter) {
+      case SelectedFilter.ACTIVE:
+        return todos.filter(todo => !todo.completed);
+
+      case SelectedFilter.COMPLETED:
+        return todos.filter(todo => todo.completed);
+
+      default:
+        return todos;
+    }
+  }, [todos, filter]);
+
   if (!todoService.USER_ID) {
     return <UserWarning />;
   }
@@ -242,72 +254,22 @@ export const App: React.FC = () => {
 
         <section className="todoapp__main" data-cy="TodoList">
           <TransitionGroup>
-            <TodoList
-              todos={todos}
-              onDelete={deleteTodo}
-              loadingIds={loadingIds}
-              filter={filter}
-              onUpdateCheckbox={updateTodoCheckbox}
-              onUpdateTitle={updateTodoTitle}
-            />
-
+            {filteredTodos.map((todo: Todo) => (
+              <CSSTransition key={todo.id} timeout={300} classNames="item">
+                <TodoItem
+                  todo={todo}
+                  onDelete={deleteTodo}
+                  loadingIds={loadingIds}
+                  onUpdateCheckbox={updateTodoCheckbox}
+                  onUpdateTitle={updateTodoTitle}
+                />
+              </CSSTransition>
+            ))}
             {creatingTodo !== null && (
               <CSSTransition key={0} timeout={300} classNames="temp-item">
                 <TodoItem todo={creatingTodo} loadingIds={loadingIds} />
               </CSSTransition>
             )}
-
-            <>
-              {/* This todo is being edited */}
-              {/* <div data-cy="Todo" className="todo">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label> */}
-              {/* This form is shown instead of the title and remove button */}
-              {/* <form>
-              <input
-                data-cy="TodoTitleField"
-                type="text"
-                className="todo__title-field"
-                placeholder="Empty todo will be deleted"
-                // value="Todo is being edited now"
-                // remove error from console
-                defaultValue="Todo is being edited now"
-              />
-            </form> */}
-              {/* <div data-cy="TodoLoader" className="modal overlay">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div> */}
-              {/* This todo is in loadind state */}
-              {/* <div data-cy="Todo" className="todo">
-            <label className="todo__status-label">
-              <input
-                data-cy="TodoStatus"
-                type="checkbox"
-                className="todo__status"
-              />
-            </label>
-
-            <span data-cy="TodoTitle" className="todo__title">
-              Todo is being saved now
-            </span>
-
-            <button type="button" className="todo__remove" data-cy="TodoDelete">
-              x
-            </button> */}
-              {/* 'is-active' class puts this modal on top of the todo */}
-              {/* <div data-cy="TodoLoader" className="modal overlay is-active">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          </div> */}
-            </>
           </TransitionGroup>
         </section>
         {/* Hide the footer if there are no todos */}
